Meta server: fix Kerberos authentication handshake in client state machine. Kfsio lib: make kereberos parameter names consistent with meta server kerberos parameter names. Fix to compile on linux with older gcc. Kerberos client meta server authentication (without ssl psk) now started working.

+++ src/cc/krb/KrbClient.cc
@@ -198,13 +198,9 @@ public:
         return 0;
     }
     int GetErrorCode() const
-    {
-        return mErrCode;
-    }
+        { return (int)mErrCode; }
     long GetLastCredEndTime() const
-    {
-        return mLastCredEndTime;
-    }
+        { return mLastCredEndTime; }
 private:
     string            mServiceHost;
     string            mKeyTabFileName;
